Remove Jansson and roll our own JSON parsing
All commands follow a pretty basic structure: an array of objects with
properties that map to numbers. We probably don't need to rely on a full
featured library to parse this JSON, especially because I imagine that
most input we see will be from trusted soures.

This parser does bend one rule in favor of ease of implementation: it
ignores all commas.