Improve saving of parameters
The watch.html page wasn't properly saving the discovered parameters to
the server, and there was no method to save the control parameters.
Fixed both of these issues.

More likely control parameters are iterated through first.

Allow users to manually enter frequency; this might be useful for
countries outside of the US.